+++ spec/views/catalog/_members_multi_volume_work.html.erb_spec.rb
@@ -10,8 +10,10 @@ RSpec.describe "catalog/_members_multi_volume_work" do
     let(:parent) { FactoryBot.create_for_repository(:scanned_resource, title: "Mui", rights_statement: "y", member_ids: [child.id]) }
     let(:document) { Valkyrie::MetadataAdapter.find(:index_solr).resource_factory.from_resource(resource: parent) }
     let(:solr_document) { SolrDocument.new(document) }
+    let(:change_set) { DynamicChangeSet.new(solr_document.resource) }
     before do
       assign :document, solr_document
+      assign :change_set, change_set
       render
     end
 
@@ -24,6 +26,9 @@ RSpec.describe "catalog/_members_multi_volume_work" do
       expect(rendered).not_to have_link href: solr_document_path(child)
       expect(rendered).to have_link "View", href: parent_solr_document_path(parent, child.id)
       expect(rendered).to have_link "Edit", href: edit_scanned_resource_path(child.id)
+      expect(rendered).to have_selector "button", text: "Detach"
+      expect(rendered).to have_selector "input#scanned_resource_member_ids"
+      expect(rendered).to have_selector "button", text: "Attach"
     end
   end
 end